Controls have names now, MapInstance sends controls to the GameMap now
Still trying to figure out where to handle keybindings. Probably going to
extend ControlMapper for it, and create a ControlCategory class to display and
group controls for a later keybind menu. You won't subclass it; you'll make an
instance, name it, and it'll register itself into ControlMapper automatically.

God damn it'll be so nice to see this work.