Made entity -> map connection, renamed an addToMap() I miseed to add()
Had to adapt Mode and the current subclassing modes to use m_controller.getMap() to get the map they're on, rather than it being passed in. 'Least now there's a reasonable guarantee that an entity is associated with only one map.

If an entity is on more than one map with this, then something really bad happened and issues are elsewhere.